feat: use forkJoin to check successful deletion of all topics

+++ apps/rxjs/11-high-order-operator-bug/src/app/app.service.ts
@@ -1,5 +1,6 @@
 import { inject, Injectable } from '@angular/core';
-import { merge, Observable, of } from 'rxjs';
+import { forkJoin, merge, Observable, of } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { LocalDBService, TopicType } from './localDB.service';
 
 @Injectable({ providedIn: 'root' })
@@ -10,10 +11,30 @@ export class AppService {
 
   deleteOldTopics(type: TopicType): Observable<boolean> {
     const infoByType = this.dbService.searchByType(type);
-    return infoByType.length > 0
-      ? infoByType
-          .map((t) => this.dbService.deleteOneTopic(t.id))
-          .reduce((acc, curr) => merge(acc, curr), of(true))
-      : of(true);
+
+    if (infoByType.length === 0) {
+      return of(true);
+    }
+
+    const deleteObservables = infoByType.map((t) =>
+      this.dbService.deleteOneTopic(t.id),
+    );
+
+    return forkJoin(deleteObservables).pipe(
+      map((results) => results.every((result) => result === true)),
+    );
+  }
+
+  // This doesnt check result directly on deleteOneTopic results.
+  // but I call searchByType again and check if all topics are deleted. (not the best way)
+  deleteOldTopics2(type: TopicType): Observable<boolean> {
+    const infoByType = this.dbService.searchByType(type);
+    if (!infoByType || infoByType.length > 0) {
+      infoByType
+        .map((t) => this.dbService.deleteOneTopic(t.id))
+        .reduce((acc, curr) => merge(acc, curr), of(true));
+    }
+
+    return this.dbService.searchByType(type).length ? of(false) : of(true);
   }
 }
